refactor(shared): extract VERSION reader into src/shared/version.ts
The same 15-line VERSION-probing loop was copy-pasted across
src/cli.ts, src/products/manifest.ts, src/products/gardener/cli.ts,
and src/meta/skill-tools/cli.ts — three of them also hard-coded their
own sourceRelativeDir. Pulls the logic into a single helper so future
product additions get version reading for free.

- New `src/shared/version.ts` with `readOwnVersion(metaUrl, sourceDir)`
  and `readPackageVersion(metaUrl, pkgName)`.
- Adds `#shared/*` package import alias.
- Rewrites all four call sites to use the shared helpers.
- `readOwnVersion` now walks upward from the caller to find the source
  tree instead of probing three hand-coded fallback paths, which is
  more robust for future layout changes.
